+++ examples/discovery-components-react-example/cypress/integration/basic/spec.ts
@@ -3,13 +3,17 @@ describe('basic test', () => {
     cy.server();
     cy.fixture('basic/query.json').as('queryJSON');
     cy.route('POST', '**/query?version=2019-01-01', '@queryJSON').as('postQuery');
-  });
-
-  it('loads page', () => {
     cy.visit('/');
-    cy.get('.bx--search input').type('abil{enter}');
-    cy.wait('@postQuery')
-      .its('requestBody.natural_language_query')
-      .should('eq', 'abil');
+  });
+  describe('When entering a query', () => {
+    beforeEach(() => {
+      cy.get('.bx--search-input').type('abil{enter}');
+    });
+
+    it('makes the appropriate query request', () => {
+      cy.wait('@postQuery')
+        .its('requestBody.natural_language_query')
+        .should('eq', 'abil');
+    });
   });
 });