+++ packages/discovery-components-react/src/components/RichPreview/RichPreview.tsx
@@ -1,0 +1,13 @@
+import React, { SFC } from 'react';
+import PdfViewer from './components/PdfViewer/PdfViewer';
+
+interface Props {
+  /**
+   * PDF file data as base64-encoded string
+   */
+  file: string;
+}
+
+const RichPreview: SFC<Props> = ({ file }) => <PdfViewer file={file} page={1} scale={1} />;
+
+export default RichPreview;