+++ packages/discovery-components-react/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.19.0](http:///discovery-components-react/compare/@disco-widgets/react-components@1.18.0...@disco-widgets/react-components@1.19.0) (2019-09-23)
+
+
+### Features
+
+* Refinements config to agg ([#118](http:///discovery-components-react/issues/118)) ([705abb0](http:///discovery-components-react/commits/705abb0))
+
+
+
+
+
+# [1.18.0](http:///discovery-components-react/compare/@disco-widgets/react-components@1.17.0...@disco-widgets/react-components@1.18.0) (2019-09-23)
+
+
+### Features
+
+* generate node sdk ([#111](http:///discovery-components-react/issues/111)) ([ca42740](http:///discovery-components-react/commits/ca42740))
+
+
+
+
+
+# [1.17.0](http:///discovery-components-react/compare/@disco-widgets/react-components@1.16.1...@disco-widgets/react-components@1.17.0) (2019-09-21)
+
+
+### Features
+
+* PDF Viewer component ([#115](http:///discovery-components-react/issues/115)) ([92effb5](http:///discovery-components-react/commits/92effb5))
+
+
+
+
+
+## [1.16.1](http:///discovery-components-react/compare/@disco-widgets/react-components@1.16.0...@disco-widgets/react-components@1.16.1) (2019-09-20)
+
+**Note:** Version bump only for package @disco-widgets/react-components
+
+
+
+
+
 # [1.16.0](http:///discovery-components-react/compare/@disco-widgets/react-components@1.15.0...@disco-widgets/react-components@1.16.0) (2019-09-20)
 
 ### Features